feat: オプション検索バーを追加 (パフォーマンス最適化済み)
